requirements.txt
langchain
langchain-community
langchain-openai
streamlit
python-dotenv
faiss-cpu
pypdf2
fastapi
uvicorn